[algorithms][envelope] merge implementation of envelope for linestrings and multilinestrings into a single file for linear geometries
